+++ esp-radio-preempt-driver/src/lib.rs
@@ -9,20 +9,14 @@
 //!
 //! ## Implementing a scheduler driver
 //!
-//! In order to hook up a scheduler, implement the `Scheduler` trait for a struct, and register it
-//! using the `scheduler_impl!()` macro. Only one scheduler can be registered in a firmware.
+//! This crate abstracts the capabilities of FreeRTOS. The scheduler must implement the following
+//! capabilities:
 //!
-//! Example:
-//!
-//! ```rust,ignore
-//! struct MyScheduler {}
-//!
-//! impl esp_preempt::Scheduler for MyScheduler {
-//!     // impl goes here
-//! }
-//!
-//! esp_preempt::scheduler_impl!(static SCHEDULER: MyScheduler = MyScheduler {});
-//! ```
+//! - A preemptive task scheduler: [`Scheduler`]
+//! - Mutexes: [`mutex::MutexImplementation`]
+//! - Semaphores: [`semaphore::SemaphoreImplementation`]
+//! - Queues: [`queue::QueueImplementation`]
+//! - Timers (functions that are executed at a specific time): [`timer::TimerImplementation`]
 //!
 //! [`esp-preempt`]: https://crates.io/crates/esp-preempt
 
@@ -45,6 +39,7 @@ unsafe extern "Rust" {
     fn esp_preempt_enable();
     fn esp_preempt_disable();
     fn esp_preempt_yield_task();
+    fn esp_preempt_yield_task_from_isr();
     fn esp_preempt_current_task() -> *mut c_void;
     fn esp_preempt_max_task_priority() -> u32;
     fn esp_preempt_task_create(
@@ -91,6 +86,12 @@ macro_rules! scheduler_impl {
         #[inline]
         fn esp_preempt_yield_task() {
             <$t as $crate::Scheduler>::yield_task(&$name)
+        }
+
+        #[unsafe(no_mangle)]
+        #[inline]
+        fn esp_preempt_yield_task_from_isr() {
+            <$t as $crate::Scheduler>::yield_task_from_isr(&$name)
         }
 
         #[unsafe(no_mangle)]
@@ -154,6 +155,73 @@ macro_rules! scheduler_impl {
 ///
 /// This trait needs to be implemented by a driver crate to integrate esp-radio with a software
 /// platform.
+///
+/// The following snippet demonstrates the boilerplate necessary to implement a scheduler using the
+/// `Scheduler` trait:
+///
+/// ```rust,no_run
+/// struct MyScheduler {}
+///
+/// impl esp_radio_preempt_driver::Scheduler for MyScheduler {
+///
+///     fn initialized(&self) -> bool {
+///         unimplemented!()
+///     }
+///
+///     fn enable(&self) {
+///         unimplemented!()
+///     }
+///
+///     fn disable(&self) {
+///         unimplemented!()
+///     }
+///
+///     fn yield_task(&self) {
+///         unimplemented!()
+///     }
+///
+///     fn yield_task_from_isr(&self) {
+///         unimplemented!()
+///     }
+///
+///     fn max_task_priority(&self) -> u32 {
+///         unimplemented!()
+///     }
+///
+///     fn task_create(
+///        &self,
+///        task: extern "C" fn(*mut c_void),
+///        param: *mut c_void,
+///        priority: u32,
+///        pin_to_core: Option<u32>,
+///        task_stack_size: usize,
+///     ) -> *mut c_void {
+///         unimplemented!()
+///     }
+///
+///     fn current_task(&self) -> *mut c_void {
+///         unimplemented!()
+///     }
+///
+///     fn schedule_task_deletion(&self, task_handle: *mut c_void) {
+///         unimplemented!()
+///     }
+///
+///     fn current_task_thread_semaphore(&self) -> SemaphorePtr {
+///         unimplemented!()
+///     }
+///
+///     fn usleep(&self, us: u32) {
+///         unimplemented!()
+///     }
+///
+///     fn now(&self) -> u64 {
+///         unimplemented!()
+///     }
+/// }
+///
+/// esp_radio_preempt_driver::scheduler_impl!(static SCHEDULER: MyScheduler = MyScheduler {});
+/// ```
 pub trait Scheduler: Send + Sync + 'static {
     /// This function is called by `esp_radio::init` to verify that the scheduler is properly set
     /// up.
@@ -165,8 +233,11 @@ pub trait Scheduler: Send + Sync + 'static {
     /// This function is called by `esp-radio` to stop the task scheduler.
     fn disable(&self);
 
-    /// This function is called by `esp_radio::init` to yield control to another task.
+    /// This function is called by `esp_radio` to yield control to another task.
     fn yield_task(&self);
+
+    /// This function is called by `esp_radio` to yield control to another task.
+    fn yield_task_from_isr(&self);
 
     /// This function is called by `esp_radio::init` to retrieve a pointer to the current task.
     fn current_task(&self) -> *mut c_void;
@@ -234,6 +305,12 @@ pub fn disable() {
 #[inline]
 pub fn yield_task() {
     unsafe { esp_preempt_yield_task() }
+}
+
+/// Yields control to another task for an interrupt.
+#[inline]
+pub fn yield_task_from_isr() {
+    unsafe { esp_preempt_yield_task_from_isr() }
 }
 
 /// Returns a pointer to the current task.